Move the FocusTracker from /wtf to /app
